+++ apps/frontend/src/app/[locale]/(main-admin-content)/managing/_components/WordCard/components/WordForms/constants.ts
@@ -9,4 +9,6 @@ export const FlagByAreaEnum = {
   ru: 'rusFlag',
   es: 'esFlag',
   fr: 'frFlag',
+  de: 'deFlag',
+  pt: 'ptFlag',
 };